dom/html/test/test_bug557620.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=557620
-->
<head>
  <title>Test for Bug 557620</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=557620">Mozilla Bug 557620</a>
<p id="display"></p>
<div id="content" style="display: none">
  <input type="tel" id='i'>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 557620 **/

// More checks are done in test_bug551670.html.

var tel = document.getElementById('i');
is(tel.type, 'tel', "input with type='tel' should return 'tel'");

</script>
</pre>
</body>
</html>
